fix(myaudio): fix precision loss and race condition in audio buffers (#1689)
* fix(myaudio): fix precision loss and race condition in audio buffers

Fix two bugs in the myaudio package:

1. Sub-second precision loss in ReadSegment (capture_buffer.go)
   - The calculation `int(offset.Seconds()) * sampleRate * bytesPerSample`
     truncated sub-second values before multiplication
   - Example: 1.5s became 1s, losing 0.5s of audio positioning
   - Fixed by multiplying in float64 before converting to int

2. Race condition in buffer pool initialization (analysis_buffer.go)
   - Multiple goroutines could race to initialize overlapSize, readSize,
     and readBufferPool globals
   - Fixed using sync.Once to ensure single initialization

Added comprehensive tests for both fixes.

* refactor(test): use strings.Contains and remove redundant reset

- Replace custom containsString/searchSubstring with strings.Contains
- Remove redundant resetAnalysisBufferGlobals() call in test

Addresses code review feedback from Gemini and CodeRabbit.

---------

+++ internal/myaudio/capture_buffer.go
@@ -463,14 +463,16 @@ func (cb *CaptureBuffer) ReadSegment(requestedStartTime time.Time, duration int)
 		startOffset := requestedStartTime.Sub(cb.startTime)
 		endOffset := requestedEndTime.Sub(cb.startTime)
 
-		startIndex := int(startOffset.Seconds()) * cb.sampleRate * cb.bytesPerSample
-		endIndex := int(endOffset.Seconds()) * cb.sampleRate * cb.bytesPerSample
+		// Calculate byte indices with full floating-point precision before truncating to int.
+		// This preserves sub-second accuracy (e.g., 1.5s * 48000 * 2 = 144000, not 96000).
+		startIndex := int(startOffset.Seconds() * float64(cb.sampleRate) * float64(cb.bytesPerSample))
+		endIndex := int(endOffset.Seconds() * float64(cb.sampleRate) * float64(cb.bytesPerSample))
 
 		startIndex %= cb.bufferSize
 		endIndex %= cb.bufferSize
 
 		if startOffset < 0 {
-			if cb.writeIndex == 0 || cb.writeIndex+int(startOffset.Seconds())*cb.sampleRate*cb.bytesPerSample > cb.bufferSize {
+			if cb.writeIndex == 0 || cb.writeIndex+int(startOffset.Seconds()*float64(cb.sampleRate)*float64(cb.bytesPerSample)) > cb.bufferSize {
 				cb.lock.Unlock()
 
 				enhancedErr := errors.Newf("requested start time is outside the buffer's current timeframe").
